Quite los botones a los que no puede acceder el administrador
